# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ariadne Metadata Management API
  version: 0.1.0
  description: >
    Phase 0 contract for Ariadne backend. Covers authentication, metadata CRUD,
    lineage management (with lineage_source), bulk import/export, semantic search,
    and AI assistant endpoints. All endpoints are prefixed with `/api/v1`.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /tables/{table_id}:
    get:
      tags: [tables]
      summary: Get table detail
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Table detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [tables]
      summary: Update table metadata
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [tables]
      summary: Delete table metadata
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
...
  /tags:
    get:
      tags: [tags]
      summary: Get tag tree (hierarchical structure)
      parameters:
        - in: query
          name: level
          schema:
            type: integer
            minimum: 1
            maximum: 3
          description: Filter by tag level
        - in: query
          name: parent_id
          schema:
            type: string
            format: uuid
          description: Get children of specific parent tag
      responses:
        '200':
          description: Tag tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagTreeResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [tags]
      summary: Create tag (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /tags/{tag_id}:
    put:
      tags: [tags]
      summary: Update tag (admin only)
      parameters:
        - $ref: '#/components/parameters/tag_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [tags]
      summary: Delete tag (admin only)
      parameters:
        - $ref: '#/components/parameters/tag_id'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete tag with children or in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/Error'
  /tags/{tag_id}/usage:
    get:
      tags: [tags]
      summary: Get tag usage statistics
      parameters:
        - $ref: '#/components/parameters/tag_id'
      responses:
        '200':
          description: Tag usage info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagUsageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}/tags:
    get:
      tags: [tags]
      summary: Get tags for a table
      parameters:
        - $ref: '#/components/parameters/table_id'
      responses:
        '200':
          description: Table tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [tags]
      summary: Add tags to table
      parameters:
        - $ref: '#/components/parameters/table_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TableTagsAddRequest'
      responses:
        '201':
          description: Tags added
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /tables/{table_id}/tags/{tag_id}:
    delete:
      tags: [tags]
      summary: Remove tag from table
      parameters:
        - $ref: '#/components/parameters/table_id'
        - $ref: '#/components/parameters/tag_id'
      responses:
        '204':
          description: Tag removed
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    table_id:
      in: path
      name: table_id
      required: true
      schema:
        type: string
        format: uuid
...
    tag_id:
      in: path
      name: tag_id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Table:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        name:
          type: string
        schema_name:
          type: string
          description: Logical schema or database namespace
        qualified_name:
          type: string
          description: Fully qualified name, e.g., schema.table
        type:
          type: string
          example: table
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        row_count:
          type: integer
        field_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TableDetail:
      allOf:
        - $ref: '#/components/schemas/Table'
      properties:
        fields:
          type: array
          items:
            $ref: '#/components/schemas/Field'
...
    TableUpdateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Field:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        name:
          type: string
        data_type:
          type: string
        description:
          type: string
        is_nullable:
          type: boolean
        is_primary_key:
          type: boolean
        is_foreign_key:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        level:
          type: integer
          minimum: 1
          maximum: 3
        path:
          type: string
          description: Full path like 'F10-公司概况-公司资料'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TagWithChildren:
      allOf:
        - $ref: '#/components/schemas/Tag'
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/TagWithChildren'
    TagCreateRequest:
      type: object
      required: [name, level]
      properties:
        name:
          type: string
        parent_id:
          type: string
          format: uuid
          nullable: true
        level:
          type: integer
          minimum: 1
          maximum: 3
    TagUpdateRequest:
      type: object
      properties:
        name:
          type: string
    TagTreeResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TagWithChildren'
    TagListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
    TagUsageResponse:
      type: object
      properties:
        tag_id:
          type: string
          format: uuid
        tag_name:
          type: string
        table_count:
          type: integer
        tables:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              source_id:
                type: string
                format: uuid
    TableTagsAddRequest:
      type: object
      required: [tag_ids]
      properties:
        tag_ids:
          type: array
          items:
            type: string
            format: uuid
...